requirements.txt
Django==4.0.5
djangorestframework==3.13.0
django-allauth==0.13.0
django-rest-auth==0.9.5
psycopg2-binary==2.9.3
pillow==8.4.0
djangorestframework-simplejwt==5.0.0
autopep8==1.6.0
autoflake==1.4
django-cors-headers==3.11.0
isort==5.10.1
drf-yasg==1.20.0
facebook-sdk==3.1.0
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.3
httplib2==0.20.4
oauthlib==3.2.0
python-twitter==3.5
tweepy==4.10.0
requests==2.28.1
idna==3.3